+++ src/components/FooterBar.js
@@ -1,13 +1,21 @@
 import '../App.css';
 import '../styles.css';
+import { Link } from 'react-router-dom';
+
 
 function FooterBar() {
     return (
       <nav className="menubar">
         <div>
-          <button>Contact</button>
-          <button>Careers</button>
-          <button>Help</button>
+          <Link to="/">
+            <button>Contact</button>
+          </Link>
+          <Link to="/">
+            <button>Careers</button>
+          </Link>
+          <Link to="/">
+            <button>Help</button>
+          </Link>
         </div>
       </nav> 
     )
