+++ src/components/UserCard.js
@@ -2,20 +2,11 @@ import { useState } from "react";
 import '../App.css';
 import '../styles.css';
 import { Link } from "react-router-dom";
+import FollowButton from "./FollowButton";
 
 function UserCard(props) {
   // Following boolean
   const [following, setFollowing] = useState(false);
-  const [followText, setFollowText] = useState("follow");
-
-  // Switch the boolean
-  const handleFollow = () => {
-    setFollowing(!following);
-    setFollowText(followText === "follow" ? "following" : "follow");
-  }
-
-  const followButtonColor = following ? "var(--color-5)" : 'var(--color-2)';
-  const followButtonTextColor = following? "var(--color-black)" : "var(--color-white)";
 
   return (
     <div className='div'>
@@ -34,15 +25,7 @@ function UserCard(props) {
         }}>
         <img src={props.profilePicture} className="user-pfp" width={100} height={100} alt="pfp" />
       </Link>
-      <div>
-        <button 
-          className="follow-button" 
-          style={{backgroundColor: followButtonColor, color: followButtonTextColor}} 
-          onClick={handleFollow}
-        >
-          {followText}
-        </button>
-      </div>
+      <FollowButton following={following} setFollowing={setFollowing}/>
     </div>
   )
 }
